+++ CHANGELOG.md
@@ -9,6 +9,12 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+### Added
+
+- Add `config.AsURL()` and `AsURLDefault()`
+
 ## [1.0.0] - 2020-12-21
 
 This is the first stable release. There have been no changes to the API since